refactor: Input, Validate 클래스 별도 파일로 분리
- App 내에서 쓸모 없어진 import 삭제, 경로 수정
